Feature/add options

+++ internal/ktea/ktea.go
@@ -18,7 +18,8 @@ var baseStyle = lipgloss.NewStyle().
 var selection string = ""
 
 type model struct {
-	table  table.Model
+	table table.Model
+
 	width  int
 	height int
 }
@@ -63,11 +64,12 @@ func center(s string, w int) string {
 	return strings.Repeat(" ", div) + s + strings.Repeat(" ", div)
 }
 
-func Ktea() {
-	files, err := os.ReadDir(os.Getenv("HOME") + "/.kube")
-	if err != nil {
-		log.Fatal(err)
-	}
+func Ktea(strFlag string, myDir string) {
+	// files, err := os.ReadDir(os.Getenv("HOME") + "/.kube")
+	// files, err := os.ReadDir(myDir)
+	// if err != nil {
+	//		log.Fatal(err)
+	//	}
 
 	centeredTitle := center("Kube Configs", 30)
 
@@ -77,7 +79,19 @@ func Ktea() {
 
 	rows := []table.Row{}
 
-	for _, file := range files {
+	dirs, err := os.ReadDir(myDir)
+	if err != nil {
+		log.Println(err)
+	}
+
+	for _, e := range dirs {
+		if e.IsDir() {
+			rows = append(rows, table.Row{e.Name()})
+		}
+	}
+
+	//	for _, file := range files {
+	for _, file := range dirs {
 		if file.Type().IsRegular() {
 			rows = append(rows, table.Row{file.Name()})
 		}
@@ -110,11 +124,20 @@ func Ktea() {
 	}
 
 	if selection != "" {
-		if _, err := os.Lstat(os.Getenv("HOME") + "/.kube/config"); err == nil {
-			os.Remove(os.Getenv("HOME") + "/.kube/config")
+		if strFlag == "env" {
+			fmt.Printf("Setting ${KUBECONFIG} to ${HOME}/.kube/%s\n", selection)
+			os.Setenv("KUBECONFIG", os.Getenv("HOME")+"/.kube"+selection)
+			fmt.Println("KUBECONFIG:", os.Getenv("KUBECONFIG"))
+		} else {
+			if _, err := os.Lstat(os.Getenv("HOME") + "/.kube/config"); err == nil {
+				os.Remove(os.Getenv("HOME") + "/.kube/config")
+			}
+			fmt.Printf("Linking ${HOME}/.kube/conifg -> %s\n", selection)
+			err := os.Symlink(os.Getenv("HOME")+"/.kube/"+selection, os.Getenv("HOME")+"/.kube/config")
+			if err != nil {
+				log.Println(err)
+			}
 		}
-		fmt.Printf("Linking ${HOME}/.kube/conifg -> %s\n", selection)
-		os.Symlink(os.Getenv("HOME")+"/.kube/"+selection, os.Getenv("HOME")+"/.kube/config")
 	} else {
 		fmt.Println("No changes made.")
 	}
